Changes to IRQ intrupt, IRQ handler, MCU changes for intrrupts, setup for function when IRQ is raised, made functions for shortcutting setting up short conversion mode and continous conversion mode for later, and some other stuff

+++ include/SPI.h
@@ -1,7 +1,7 @@
 #include <stdint.h>
 
 void SPI_SendCommand(uint8_t cmd);
-void SPI_ReceiveData(uint8_t *buffer, uint16_t size);
+void SPI_ReceiveData(uint8_t *buffer, uint8_t size);
 void PULL_ADC_CS_LOW(void);
 void PULL_ADC_CS_HIGH(void);
 void PULL_DAC_CS_LOW(void);
